basic refactoring of all helper widgets
