+++ Mobile-Assessment-Flutter-staging/lib/screens/congratulations.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:mobile_assessment_flutter/background_wrapper.dart';
-import 'package:mobile_assessment_flutter/screens/custom_registration_screen.dart';
-import 'package:mobile_assessment_flutter/screens/welcome_screen.dart';
+import 'package:mobile_assessment_flutter/widgets/background_wrapper.dart';
+import 'package:mobile_assessment_flutter/widgets/custom_button.dart';
+import 'package:mobile_assessment_flutter/widgets/header_text.dart';
 
 class CongratulationScreen extends StatelessWidget {
   static const String id = 'Congratulation_screen';
@@ -16,42 +16,50 @@ class CongratulationScreen extends StatelessWidget {
         Scaffold(
           backgroundColor: Colors.transparent,
           body: Padding(
-              padding: const EdgeInsets.only(top: 110, right: 15, left: 15),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  const TextHeader(
-                      header: 'Congratulations!',
-                      subheader:
-                          'Your account as been successfully \ncreated.'),
-                  Padding(
-                    padding: const EdgeInsets.only(
-                        left: 40, right: 40, top: 25, bottom: 60),
-                    child: SizedBox(
-                      height: size.height * .4,
-                      width: size.width,
-                      child: const Image(
-                        image: AssetImage(
-                          'assets/ic-congratulations.png',
-                        ),
-                        fit: BoxFit.fill,
+            padding: const EdgeInsets.only(top: 110, right: 15, left: 15),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                const TextHeader(
+                  header: 'Congratulations!',
+                  subheader: 'Your account as been successfully \ncreated.',
+                ),
+                Padding(
+                  padding: const EdgeInsets.only(
+                      left: 40, right: 40, top: 25, bottom: 60),
+                  child: SizedBox(
+                    height: size.height * .4,
+                    width: size.width,
+                    child: const Image(
+                      image: AssetImage(
+                        'assets/ic-congratulations.png',
                       ),
+                      fit: BoxFit.fill,
                     ),
                   ),
-                  Center(
-                    child: CustomButton(
-                        text: const Text(
-                          'Continue',
-                          style: TextStyle(
-                              fontSize: 25, fontWeight: FontWeight.w700),
-                        ),
-                        onPressed: () {
-                          // Navigator.pushNamed(context, .id);
-                        },
-                        size: Size(size.width * .5, size.height * .08)),
-                  )
-                ],
-              )),
+                ),
+                Center(
+                  child: CustomButton(
+                    text: const Text(
+                      'Continue',
+                      style: TextStyle(
+                        fontSize: 20,
+                        fontWeight: FontWeight.w600,
+                        color: Colors.white70,
+                      ),
+                    ),
+                    onPressed: () {
+                      // Navigator.pushNamed(context, .id);
+                    },
+                    size: Size(
+                      size.width * .5,
+                      size.height * .08,
+                    ),
+                  ),
+                )
+              ],
+            ),
+          ),
         )
       ],
     );
