+++ Mobile-Assessment-Flutter-staging/lib/widgets/custom_button.dart
@@ -1,0 +1,27 @@
+import 'package:flutter/material.dart';
+
+class CustomButton extends StatelessWidget {
+  final Widget text;
+  final VoidCallback onPressed;
+  final Size size;
+  const CustomButton({
+    Key? key,
+    required this.text,
+    required this.onPressed,
+    required this.size,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return ElevatedButton(
+        onPressed: onPressed,
+        style: ElevatedButton.styleFrom(
+          // primary: Colors.white,
+          minimumSize: size,
+          shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(15),
+          ),
+        ),
+        child: text);
+  }
+}
