+++ Mobile-Assessment-Flutter-staging/lib/widgets/otp_input.dart
@@ -1,0 +1,35 @@
+import 'package:flutter/material.dart';
+
+class OtpInput extends StatelessWidget {
+  final TextEditingController controller;
+  const OtpInput(this.controller, {Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      decoration: BoxDecoration(
+        borderRadius: BorderRadius.circular(15),
+        color: Colors.white,
+      ),
+      height: 55,
+      width: 50,
+      child: TextField(
+        autofocus: false,
+        textAlign: TextAlign.center,
+        keyboardType: TextInputType.number,
+        controller: controller,
+        maxLength: 1,
+        cursorColor: Theme.of(context).primaryColor,
+        decoration: const InputDecoration(
+            border: InputBorder.none,
+            counterText: '',
+            hintStyle: TextStyle(color: Colors.black, fontSize: 20.0)),
+        onChanged: (value) {
+          if (value.length == 1) {
+            FocusScope.of(context).nextFocus();
+          }
+        },
+      ),
+    );
+  }
+}
