+++ Mobile-Assessment-Flutter-staging/lib/widgets/text_field_container.dart
@@ -1,0 +1,28 @@
+import 'package:flutter/material.dart';
+
+class TextFieldContainer extends StatelessWidget {
+  final Widget child;
+  final double? height;
+  final Color? color;
+  const TextFieldContainer({
+    Key? key,
+    required this.child,
+    this.height,
+    this.color,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    Size size = MediaQuery.of(context).size;
+    return Container(
+      height: height ?? size.height * 0.065,
+      width: size.width,
+      decoration: BoxDecoration(
+        color: color ?? Colors.white,
+        borderRadius: BorderRadius.circular(15),
+        shape: BoxShape.rectangle,
+      ),
+      child: child,
+    );
+  }
+}
